+++ src/lib/input-validation/index.ts
@@ -45,3 +45,4 @@ export * from "./validateRegex";
 export * from "./validateJSON";
 export * from "./constant/errorType";
 export * from "./constant/numbers";
+export * from "./utils/regexGenerator";