+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.25.0]
+### Added
+- Added regex generator utils
+
+### Changed
+- Allow validateRegex to validate unicode character
+
 ## [1.24.1]
 ### Fixed
 - Fixed `validateDomain` regex constructor need to be double escaped